Refactor scene source editors into separate files
Extract per-source property editors out of SourcePropertiesPanel into three new files: SceneDeviceEditors.kt, SceneGraphicEditors.kt and SceneMediaEditors.kt. Cleans up imports and leaves SourcePropertiesPanel as the host/orchestrator. Also stabilises a screenshot flake by adding freezeCaret() and calling it in PicturesTabScreenshotTest to pin the blinking text-caret phase during captures.
